pulled out classification tests into own file. added ttest1 and bernstein tests.
