+++ tests/benchmark_tools_test.py
@@ -6,29 +6,10 @@ from string import ascii_letters
 import numpy as np
 import pandas as pd
 import scipy.stats as ss
-from sklearn.metrics import brier_score_loss, log_loss, zero_one_loss
 import benchmark_tools.benchmark_tools as bt
 import benchmark_tools.classification as btc
 from benchmark_tools import util
 from benchmark_tools.test_constants import MC_REPEATS_LARGE, FPR
-
-
-def hard_loss_binary(y_bool, log_pred_prob, FP_cost=1.0):
-    '''Special case of hard_loss.'''
-    N, n_labels = btc.shape_and_validate(y_bool, log_pred_prob)
-    assert(n_labels == 2)
-    assert(FP_cost > 0.0)
-
-    FN_cost = 1.0
-    thold = np.log(FP_cost / (FP_cost + FN_cost))
-
-    y_bool = y_bool.astype(bool)  # So we can use ~
-    yhat = log_pred_prob[:, 1] >= thold
-    assert(y_bool.dtype.kind == 'b' and yhat.dtype.kind == 'b')
-
-    loss = (~y_bool * yhat) * FP_cost + (y_bool * ~yhat) * FN_cost
-    assert(np.all((loss == 0) | (loss == FN_cost) | (loss == FP_cost)))
-    return loss
 
 
 def fp_rnd():
@@ -89,10 +70,48 @@ def test_clip_EB(runs=100):
             assert(np.allclose(np.fmin(upper, mu + EB0),
                                np.fmin(upper, mu + EB)))
 
-# TODO include test of ttest1
-#     make sure always nan for non-finite
-#     test both nan on zero and not
-#     test scales that produce underflow
+
+def test_ttest1():
+    N = np.random.randint(low=1, high=10)
+    x = np.random.randn(N)
+    x = np.random.choice(np.concatenate(([0], x)), size=N, replace=True)
+
+    all_same = np.all(x[0] == x)
+
+    pval0 = bt.ttest1(x, nan_on_zero=False)
+    pval1 = bt.ttest1(x, nan_on_zero=True)
+
+    if all_same:
+        if N <= 1 or x[0] == 0:
+            assert(pval0 == 1.0)
+        else:
+            # Not guaranteed to cause problem in scipy ttest which over-rides
+            # to make p-value 0
+            assert(np.allclose(pval0, 0.0))
+        assert(np.isnan(pval1))
+    else:
+        assert(pval0 == pval1)
+
+        if N <= 1:
+            assert(pval0 == 1.0)
+        else:
+            _, pval_ss = ss.ttest_1samp(x, 0.0)
+            assert(pval0 == pval_ss)
+            assert(0.0 < pval0 and pval0 < 1.0)
+            mu = bt.t_EB(x, confidence=1.0 - pval0)
+            assert(np.allclose(mu, np.abs(np.mean(x))))
+
+    # Now make sure infs work
+    if N >= 1:
+        x[0] = np.random.choice([-np.inf, np.inf])
+        pval0 = bt.ttest1(x, nan_on_zero=False)
+        pval1 = bt.ttest1(x, nan_on_zero=True)
+        assert(pval0 == 1.0)
+        all_same = N == 0 or np.all(x[0] == x)
+        if all_same:
+            assert(np.isnan(pval1))
+        else:
+            assert(pval1 == 1.0)
 
 
 def test_t_EB(runs=10, trials=100):
@@ -119,12 +138,41 @@ def test_t_EB(runs=10, trials=100):
     _, pval_agg = ss.combine_pvalues(pval)
     return pval_agg
 
+
+def test_bernstein_EB(runs=10, trials=100):
+    pval = []
+    while len(pval) < runs:
+        N = np.random.randint(low=0, high=10)
+        confidence = np.random.rand()
+
+        lower = np.random.randn()
+        upper = lower + np.abs(np.random.randn())
+
+        if N <= 1:
+            x = np.random.uniform(lower, upper, size=N)
+            EB = bt.bernstein_EB(x, lower, upper, confidence=confidence)
+            assert(np.allclose(EB, upper - lower))
+        else:
+            fail = 0
+            for tt in range(trials):
+                # Crank up N to test this bound
+                x = np.random.uniform(lower, upper, size=100 * N)
+                true_mu = (lower + upper) / 2
+
+                EB = bt.bernstein_EB(x, lower, upper, confidence=confidence)
+                mu = np.mean(x)
+                LB, UB = mu - EB, mu + EB
+                assert(np.isfinite(LB) and np.isfinite(UB))
+                fail += (true_mu < LB) or (UB < true_mu)
+            pval.append(ss.binom_test(fail, trials, 1.0 - confidence,
+                                      alternative='greater'))
+    _, pval_agg = ss.combine_pvalues(pval)
+    return pval_agg
+
+
 # TODO test other forms of EB
 # will need to use one sided test for bernstein
 # maybe put in slow_tests
-
-
-
 
 def test_get_mean_and_EB(runs=10, trials=100):
     # TODO test bernstein and boot as well, will need one sided
@@ -172,112 +220,6 @@ def test_get_mean_and_EB(runs=10, trials=100):
             pval.append(ss.binom_test(fail, trials, 1.0 - confidence))
     _, pval_agg = ss.combine_pvalues(pval)
     return pval_agg
-
-# TODO pull these guys out into test classification
-
-
-def hard_loss_binary_test():
-    '''Also tests hard loss.'''
-    n_labels = 2
-    N = np.random.randint(low=1, high=10)
-
-    y_bool = np.random.rand(N) <= 0.5
-    y_pred = util.normalize(np.random.randn(N, n_labels))
-    loss = hard_loss_binary(y_bool, y_pred)
-
-    act = btc.hard_loss_decision(y_pred, 1.0 - np.eye(n_labels))
-    loss2 = zero_one_loss(y_bool.astype(int), act)
-    assert(np.allclose(np.mean(loss), loss2))
-
-    loss2 = btc.hard_loss(y_bool, y_pred)
-    assert(np.allclose(loss, loss2))
-
-
-def hard_loss_decision_test():
-    n_labels = np.random.randint(low=1, high=10)
-    n_act = np.random.randint(low=1, high=10)
-    N = np.random.randint(low=0, high=10)
-
-    y_pred = util.normalize(np.random.randn(N, n_labels))
-
-    act = btc.hard_loss_decision(y_pred, 1.0 - np.eye(n_labels))
-    assert(np.all(np.argmax(y_pred, axis=1) == act))
-
-    loss_mat = np.random.rand(n_labels, n_act)
-    act = btc.hard_loss_decision(y_pred, loss_mat)
-
-    loss_mat = np.concatenate((loss_mat, np.ones((n_labels, 1))), axis=1)
-    act2 = btc.hard_loss_decision(y_pred, loss_mat)
-    assert(np.all(act == act2))
-
-    loss_mat = np.concatenate((loss_mat, np.zeros((n_labels, 1))), axis=1)
-    act2 = btc.hard_loss_decision(y_pred, loss_mat)
-    assert(np.all(act2 == loss_mat.shape[1] - 1))
-
-
-def log_loss_test():
-    n_labels = np.random.randint(low=1, high=10)
-    N = np.random.randint(low=1, high=10)
-
-    y = np.random.randint(low=0, high=n_labels, size=N)
-    y_pred = util.normalize(np.random.randn(N, n_labels))
-
-    if n_labels >= 2:
-        loss = btc.log_loss(y, y_pred)
-
-        loss2 = log_loss(y, np.exp(y_pred), labels=range(n_labels))
-        assert(np.allclose(np.mean(loss), loss2))
-
-    with np.errstate(invalid='ignore', divide='ignore'):
-        pred = np.log(util.one_hot(y, n_labels))
-    loss2 = btc.log_loss(y, pred)
-    assert(np.max(np.abs(loss2)) <= 1e-8)
-
-
-def brier_loss_test():
-    n_labels = np.random.randint(low=1, high=4)
-    N = np.random.randint(low=1, high=10)
-
-    y = np.random.randint(low=0, high=n_labels, size=N)
-    y_pred = util.normalize(np.random.randn(N, n_labels))
-
-    loss = btc.brier_loss(y, y_pred, rescale=False)
-    # sklearn learn is dumb and gets confused when only one class passed in
-    if n_labels == 2 and np.std(y) >= 1e-8:
-        loss2 = brier_score_loss(y == 1, np.exp(y_pred[:, 1]), pos_label=True)
-        assert(np.allclose(np.mean(loss), 2.0 * loss2))
-
-    with np.errstate(invalid='ignore', divide='ignore'):
-        pred = np.log(util.one_hot(y, n_labels))
-    loss2 = btc.brier_loss(y, pred, rescale=False)
-    assert(np.max(np.abs(loss2)) <= 1e-8)
-
-
-def spherical_loss_test():
-    n_labels = np.random.randint(low=1, high=10)
-    N = np.random.randint(low=1, high=10)
-
-    y = np.random.randint(low=0, high=n_labels, size=N)
-    y_pred = util.normalize(np.random.randn(N, n_labels))
-
-    loss = btc.spherical_loss(y, y_pred, rescale=False)
-
-    # Check against the linear implementation
-    pred_prob = np.exp(y_pred)
-    normalizer = np.sqrt(np.sum(pred_prob ** 2, axis=1))
-    loss_0 = -pred_prob[np.arange(N), y.astype(int)] / normalizer
-    assert(np.allclose(loss, loss_0, equal_nan=True))
-
-    with np.errstate(invalid='ignore', divide='ignore'):
-        pred = np.log(util.one_hot(y, n_labels))
-    loss2 = btc.spherical_loss(y, pred, rescale=False)
-    assert(np.max(np.abs(loss2 + 1.0)) <= 1e-8)
-
-    if n_labels >= 2:
-        with np.errstate(invalid='ignore', divide='ignore'):
-            pred_prob = util.normalize(np.log(1.0 - util.one_hot(y, n_labels)))
-        loss2 = btc.spherical_loss(y, pred_prob, rescale=False)
-        assert(np.max(np.abs(loss2)) <= 1e-8)
 
 
 def loss_summary_table_test():
@@ -335,18 +277,13 @@ if __name__ == '__main__':
 
     for _ in range(MC_REPEATS_LARGE):
         test_clip_EB()
-        hard_loss_binary_test()
-        hard_loss_decision_test()
-        log_loss_test()
-        brier_loss_test()
-        spherical_loss_test()
+        test_ttest1()
         loss_summary_table_test()
 
     print('Now running MC tests')
-    pval = test_t_EB(trials=MC_REPEATS_LARGE)
-    print(pval)
-    assert(pval >= FPR / 2.0)
-    pval = test_get_mean_and_EB(trials=MC_REPEATS_LARGE)
-    print(pval)
-    assert(pval >= FPR / 2.0)
+    test_list = [test_t_EB, test_bernstein_EB, test_get_mean_and_EB]
+    for test_f in test_list:
+        pval = test_f(trials=MC_REPEATS_LARGE)
+        print(pval)
+        assert(pval >= FPR / len(test_list))
     print('passed')
